+++ NEWS.md
@@ -8,6 +8,10 @@
 
 1. Supplying `keep.names=` to `as.integer64()` now emits a warning, which will become an error in the next release.
 
+## NEW FEATURES
+
+1. `bitwNot`, `bitwAnd`, `bitwOr`, `bitwXor`, `bitwShiftL` and `bitwShiftR` get an overload to work correctly with `integer64` (#33).
+
 ## NOTES
 
 1. The R version dependency has been bumped from 3.5.0 (2018) to 3.6.0 (2019).